src/frontend_service/requirements.txt
# Additional requirements for frontend service
streamlit-chat==0.1.1
streamlit-extras==0.3.5